requirements.txt
Django
django-cors-headers
django-extensions
PyMuPDF
PyPDF2
nltk
python-dotenv
boto3
txtai
pillow
pdf2image
pytesseract
openai
python-docx
langdetect
fitz
moviepy

